Cleaning up iOS Code
