+++ ios/FotoPresenter/uilayer/Font.swift
@@ -1,0 +1,23 @@
+//
+//  Font.swift
+//  FotoPresenter
+//
+//  Created by Kevin Schildhorn on 12/16/22.
+//
+
+import Foundation
+import SwiftUI
+
+struct Test {
+    static let largeTitle: Font = Font.custom("", size: 16.0, relativeTo: .title)
+    static let title: Font = Font.custom("", size: 16.0, relativeTo: .title)
+    static let title2: Font = Font.custom("", size: 16.0, relativeTo: .title)
+    static let title3: Font = Font.custom("", size: 16.0, relativeTo: .title)
+    static let headline: Font = Font.custom("", size: 16.0, relativeTo: .title)
+    static let subheadline: Font = Font.custom("", size: 16.0, relativeTo: .title)
+    static let body: Font = Font.custom("", size: 16.0, relativeTo: .title)
+    static let callout: Font = Font.custom("", size: 16.0, relativeTo: .title)
+    static let caption: Font = Font.custom("", size: 16.0, relativeTo: .title)
+    static let caption2: Font = Font.custom("", size: 16.0, relativeTo: .title)
+    static let footnote: Font = Font.custom("", size: 16.0, relativeTo: .title)
+}
